Revert "Feature/sos r16 visualizar menu charolas"
