--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,79 +1,0 @@
-## Plantilla PR BackEnd
-Última actualización 28/04/25
-
----
-
-# Descripción
-
-Incluye un resumen del cambio y el problema que se ha solucionado. También proporciona el contexto relevante y la motivación para este cambio. Lista cualquier dependencia requerida para esta modificación.
-
----
-
-## Tipo de cambio
-
-- [ ] Corrección de error (fix sin romper funcionalidades existentes)
-- [ ] Nueva funcionalidad (feature sin romper funcionalidades existentes)
-- [ ] Cambio disruptivo (rompe compatibilidad o modifica comportamiento actual )
-- [ ] Actualización de documentación requerida
-- [ ] Cambio mínimo (cambio visual o estructural que no afecta la lógica del sistema)
-
----
-
-# ¿Qué archivo fue el que modifique?
-Ejemplo:
-- UserController.js
-- api.js
-
----
-
-# ¿Cómo se ha probado?
-Describe berevemente cómo se probó esta funcionalidad. Ejemplo:
-- "Se probaron los endpoints '/api/login' y '/api/profile' con Postman. Las respuestas fueron correctas y sin errores en consola."
-- "Se realizaron pruebas visuales/manuales. No se detectaron impactos en lógica ni funcionalidad."
-
----
-
-# Notas para cambios menores
-- [ ] Este PR realiza un cambio mínimo que no afecta la lógica del sistema.
-- [ ] Se validó el comportamiento básico y no se detectaron efectos colaterales.
-- [ ] No se realizaron pruebas automatizadas porque no aplica
-
----
-
-# Lista de verificación del autor
-
-- [ ] El código sigue las normas de estilo del proyecto
-- [ ] He realizado una autoevaluación del código
-- [ ] El código esta comentado en las secciones complejas o no obvias
-- [ ] Documentación actualizada aplica
-- [ ] El código no genera nuevas advertencias o errores
-- [ ] Se añadieron pruebas relevantes
-
----
-
-# Lista de Verificación de Pruebas
-- [ ] Las pruebas unitarias nuevas y existentes pasan correctamente con mis cambios
-- [ ] Las pruebas de seguridad nuevas y existentes pasan correctamente con mis cambios
-- [ ] Las pruebas de regresión nuevas y existentes pasan correctamente con mis cambios
-- [ ] Las pruebas de integración nuevas y existentes pasan correctamente con mis cambios
-- [ ] Las pruebas heurísticas nuevas y existentes pasan correctamente con mis cambios
-
----
-
-## Checklist del evaluador:
-
-- [ ] La descripción del Pull Request es clara y específica
-- [ ] No se introducen errores ni inconsistencias
-- [ ] Ortografía y gramática correctas en documentación
-- [ ] El código es entendible y cumple con los estándares
-- [ ] Notifiqué al autor del PR si encontré errores, dudas o sugerencias
-- [ ] (Si se aprueba) Procedo a eliminar la rama correspondiente
-
----
-
-# Versión: V1
-### Autores
-
-| Nombre                         | Rol   |
-| ------------------------------ | ----- |
-| Juan Antonio Landeros Velázquez          | Autor |
