--- controllers/charola.controller.js
@@ -1,3 +1,0 @@
-const Charola = require("../models/charola.model.js");
-
-exports.registrarCharola = async (req, res) => {};
