+++ controllers/test.controller.js
@@ -1,0 +1,2 @@
+const dotenv = require("dotenv");
+dotenv.config();
