--- models/charola.model.js
@@ -1,18 +1,0 @@
-const db = require("../utils/database");
-
-// TODO: Revisar el modelo, ya que probablemente no es correcto porque se debe trabajar con mas tablas
-module.exports = class Charola {
-  static async registrarCharola(data) {
-    try {
-      const connection = await db();
-      const result = await connection.query(
-        "INSERT INTO charolas (nombre, fecha, alimentacion, peso, hidratacion) VALUES (?, ?, ?, ?)",
-        [data.id, data.nombre, data.descripcion, data.fecha]
-      );
-      return result;
-    } catch (error) {
-      console.error("Error al registrar la charola:", error);
-      throw error;
-    }
-  }
-};
