+++ package.json
@@ -5,8 +5,7 @@
   "main": "index.js",
   "type": "commonjs",
   "scripts": {
-    "test": "jest --verbose",
-    "type": "module"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
@@ -19,7 +18,6 @@
   },
   "homepage": "https://github.com/slugcraft/TECH-NEBRIOS-BACKEND#readme",
   "dependencies": {
-    "@prisma/client": "^6.7.0",
     "aws-sdk": "^2.1692.0",
     "bcryptjs": "^3.0.2",
     "body-parser": "^2.2.0",
@@ -31,10 +29,5 @@
     "mariadb": "^3.4.0",
     "mysql2": "^3.14.0",
     "nodejs": "^0.0.0"
-  },
-  "devDependencies": {
-    "jest": "^29.7.0",
-    "jsdoc": "^4.0.4",
-    "prisma": "^6.7.0"
   }
 }
