--- routes/charola.routes.js
@@ -1,28 +1,0 @@
-/* Dependencias */
-const express = require("express");
-const router = express.Router();
-/* ------------ */
-
-router.use(express.json());
-router.use(express.urlencoded({ extended: true }));
-
-/* INCLUDE CONTROLLERS */
-const controlador = require("../controllers/charola.controller");
-/* ------------------- */
-
-/* GET METHODS */
-/* ----------- */
-
-/* POST METHODS */
-router.post("/registrarCharola", controlador.registrarCharola);
-/* ----------- */
-
-/* PUT METHODS */
-
-/* ----------- */
-
-/* DELETE METHODS */
-
-/* ----------- */
-
-module.exports = router;
