+++ routes/test.routes.js
@@ -1,0 +1,29 @@
+/* Dependencias */
+const express = require("express");
+const bodyParser = require("body-parser");
+const router = express.Router();
+
+router.use(bodyParser.urlencoded({ extended: true }));
+router.use(express.json()); 
+router.use(express.urlencoded({ extended: true }));
+
+/* INCLUDE CONTROLLERS */
+/* ------------------- */
+
+/* GET METHODS */
+
+/* ----------- */
+
+/* POST METHODS */
+
+/* ----------- */
+
+/* PUT METHODS */
+
+/* ----------- */
+
+/* DELETE METHODS */
+
+/* ----------- */
+
+module.exports = router;
