+++ server.js
@@ -1,60 +1,40 @@
-/* Dependencias */
-const express = require("express");
-const path = require("path");
+const express = require('express');
+const path = require('path');
 const bodyParser = require("body-parser");
+const dotenv = require("dotenv");
 const session = require("express-session");
-const db = require("./utils/database");
-
 const app = express();
 
-const sesionActiva = require("./utils/middlewares/sesionActiva");
-/* ------------ */
+const db = require('./utils/database');
 
-
-/* Rutas */
-const charolasRoutes = require('./routes/menu_charolas.route');
-const usuarioRoutes = require("./routes/usuario.routes");
-/* ----- */
-
-/* Rutas de la API */
-app.use("/charola", charolasRoutes);
-app.use("/usuario", usuarioRoutes);
-/* ----- */
-
-/* Conexion a la base de datos*/
-async function testDB() {
-  try {
-    const test = await db();
-    console.log("Conexión exitosa. Resultado:", test);
-  } catch (error) {
-    console.error("Error al conectar con la base de datos:", error);
-  }
+async function testDB(){
+    try {
+        const test = await db();
+        console.log('Conexión exitosa. Resultado:', test);
+    } catch (error) {
+        console.error('Error al conectar con la base de datos:', error);
+    }
 }
 
 testDB();
-/* ----- */
 
-/* Inicializar express */
-app.set("view engine", "ejs");
+app.set('view engine', 'ejs');
 app.set("views", path.join(__dirname, "views"));
 app.use(express.static(path.join(__dirname, "public")));
-app.use(express.json()); 
 
-/* ----- */
-
-/* Route 404 Error */
+// Route 404 Error
 app.use((req, res, next) => {
-  res.status(404).render("404", { title: "404 - Not Found" });
+    res.status(404).render("404", { title: "404 - Not Found" });
 });
 app.use((req, res) => {
-  console.log("invalid url");
-  res.redirect("/");
+    console.log("invalid url");
+    res.redirect("/");
 });
-/* ----- */
-
-/* Iniciar server */
+  
+/* Start server */
 const port = process.env.PORT || 3000;
 app.listen(port, () => {
-  console.log(`Servidor corriendo en http://localhost:${port}`);
+console.log(`Servidor corriendo en http://localhost:${port}`);
 });
-/* -------- */
+  
+
